[gym/common] Fix quantity eval mode handling.
